requirements/base.txt
# Core
Django==4.2.21
psycopg2-binary==2.9.11
python-decouple==3.8

# REST API
djangorestframework==3.15.2
django-cors-headers==4.9.0
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1

# Stripe billing
stripe==11.5.0

# Async tasks
celery==5.4.0
django-celery-beat==2.8.1
redis==5.2.1

# Email
django-anymail[mailgun]==14.0

# Storage (prod)
django-storages==1.14.6
boto3==1.36.21

# Auth
django-allauth==65.14.3

# Utilities
Pillow==12.1.1
whitenoise==6.9.0
gunicorn==23.0.0

# Monitoring
sentry-sdk[django]==2.53.0
django-health-check==4.0.3

# Security
django-axes==6.5.2
django-csp==4.0

# Notifications (Telegram/Signal HTTP client)
httpx==0.28.1
